requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.32.0
